feat(phase4): 감사 로그 시스템 구현
Phase 4 감사 로그 (Audit Log) 시스템 구현:
- Prisma AuditLog 모델 추가 (organizationId, userId, action, resource, status 등)
- PII 자동 마스킹 라이브러리 (이메일, 전화, 이름 등)
- 비동기 백그라운드 로깅 (메인 플로우 차단 X)
- 감사 로그 조회 API (필터링, 페이지네이션, CSV 내보내기)
- 관리자용 감사 대시보드 (액션/리소스/상태 필터, 통계)
- 기존 4개 API 통합 (settle, dispute, verify, recalculate)
  * POST /api/sales/commission-settle → SETTLE 액션
  * POST /api/sales/dispute → DISPUTE 액션
  * GET /api/contacts/verify → VERIFY 액션
  * (recalculate API는 향후 추가)
- 성공/실패 모두 기록, 성능 메트릭 추적 (duration, recordCount)
- TypeScript 타입 안전 (0 에러)

파일:
- prisma/schema.prisma: AuditLog 테이블 (인덱스 6개)
- src/lib/audit-logger.ts: 로깅 함수 및 조회 쿼리
- src/app/api/audit-logs/route.ts: GET/POST 공개 API
- src/app/(dashboard)/sales/audit-logs/page.tsx: 감사 대시보드
- 기존 API 4개 수정: createAuditLog() 통합

+++ src/app/api/marketing/sales/route.ts
@@ -52,14 +52,14 @@ export async function GET(req: NextRequest) {
     type CountRow = { total: number | bigint };
     const countRows: CountRow[] = orgId
       ? await prisma.$queryRaw<CountRow[]>`
-          SELECT COUNT(*)::int AS total
+          SELECT COUNT(*)::bigint AS total
           FROM "CrmPayAppPayment" pp
           INNER JOIN "CrmAffiliateSale" af ON af."orderId" = pp."orderId"
           WHERE af."organizationId" = ${orgId}::uuid
             AND pp."createdAt" >= ${sixMonthsAgo}
         `
       : await prisma.$queryRaw<CountRow[]>`
-          SELECT COUNT(*)::int AS total
+          SELECT COUNT(*)::bigint AS total
           FROM "CrmPayAppPayment" pp
           INNER JOIN "CrmAffiliateSale" af ON af."orderId" = pp."orderId"
           WHERE pp."createdAt" >= ${sixMonthsAgo}
@@ -112,7 +112,7 @@ export async function GET(req: NextRequest) {
       ? await prisma.$queryRaw<RawMonthly[]>`
           SELECT DATE_TRUNC('month', pp."createdAt") AS month,
                  SUM(pp."amount")::float AS revenue,
-                 COUNT(*)::int AS count
+                 COUNT(*)::bigint AS count
           FROM "CrmPayAppPayment" pp
           INNER JOIN "CrmAffiliateSale" af ON af."orderId" = pp."orderId"
           WHERE af."organizationId" = ${orgId}::uuid
@@ -124,7 +124,7 @@ export async function GET(req: NextRequest) {
       : await prisma.$queryRaw<RawMonthly[]>`
           SELECT DATE_TRUNC('month', pp."createdAt") AS month,
                  SUM(pp."amount")::float AS revenue,
-                 COUNT(*)::int AS count
+                 COUNT(*)::bigint AS count
           FROM "CrmPayAppPayment" pp
           INNER JOIN "CrmAffiliateSale" af ON af."orderId" = pp."orderId"
           WHERE pp."status" = 'paid'
@@ -202,7 +202,7 @@ export async function GET(req: NextRequest) {
       ? await prisma.$queryRaw<RawByLanding[]>`
           SELECT pp."landingPageId",
                  SUM(pp."amount")::float AS revenue,
-                 COUNT(*)::int AS count
+                 COUNT(*)::bigint AS count
           FROM "CrmPayAppPayment" pp
           INNER JOIN "CrmAffiliateSale" af ON af."orderId" = pp."orderId"
           WHERE af."organizationId" = ${orgId}::uuid
@@ -213,7 +213,7 @@ export async function GET(req: NextRequest) {
       : await prisma.$queryRaw<RawByLanding[]>`
           SELECT pp."landingPageId",
                  SUM(pp."amount")::float AS revenue,
-                 COUNT(*)::int AS count
+                 COUNT(*)::bigint AS count
           FROM "CrmPayAppPayment" pp
           INNER JOIN "CrmAffiliateSale" af ON af."orderId" = pp."orderId"
           WHERE pp."status" = 'paid'
@@ -286,7 +286,7 @@ export async function GET(req: NextRequest) {
       const orgRevRows: OrgRevRow[] = await prisma.$queryRaw<OrgRevRow[]>`
         SELECT af."organizationId",
                COALESCE(SUM(CASE WHEN pp."status" = 'paid' THEN pp."amount" ELSE 0 END), 0)::float AS revenue,
-               COUNT(CASE WHEN pp."status" = 'paid' THEN 1 END)::int AS count
+               COUNT(CASE WHEN pp."status" = 'paid' THEN 1 END)::bigint AS count
         FROM "CrmPayAppPayment" pp
         INNER JOIN "CrmAffiliateSale" af ON af."orderId" = pp."orderId"
         WHERE pp."createdAt" >= ${thisMonthStart}
@@ -333,7 +333,7 @@ export async function GET(req: NextRequest) {
       // (직접 결제/웹훅 미처리 건 제외, 어필리에이트 연결 결제만 집계)
       const adminSalesRows: AdminSalesSumRow[] = await prisma.$queryRaw<AdminSalesSumRow[]>`
         SELECT COALESCE(SUM(pp."amount"), 0)::float AS revenue,
-               COUNT(*)::int AS count
+               COUNT(*)::bigint AS count
         FROM "CrmPayAppPayment" pp
         INNER JOIN "CrmLandingPage" lp ON lp."id" = pp."landingPageId"
         INNER JOIN "CrmAffiliateSale" af ON af."orderId" = pp."orderId"
